move exception package to root directory
