+++ src/main/java/com/epam/taskgym/exception/FailAuthenticateException.java
@@ -1,4 +1,4 @@
-package com.epam.taskgym.service.exception;
+package com.epam.taskgym.exception;
 
 public class FailAuthenticateException extends RuntimeException{
     public FailAuthenticateException(String message) {
