+++ src/main/java/com/epam/taskgym/exception/MissingAttributes.java
@@ -1,4 +1,4 @@
-package com.epam.taskgym.service.exception;
+package com.epam.taskgym.exception;
 
 public class MissingAttributes extends RuntimeException {
     public MissingAttributes(String message) {
