+++ src/main/java/com/epam/taskgym/service/TrainingService.java
@@ -5,8 +5,8 @@ import com.epam.taskgym.entity.Trainer;
 import com.epam.taskgym.entity.Training;
 import com.epam.taskgym.entity.TrainingType;
 import com.epam.taskgym.repository.TrainingRepository;
-import com.epam.taskgym.service.exception.BadRequestException;
-import com.epam.taskgym.service.exception.MissingAttributes;
+import com.epam.taskgym.exception.BadRequestException;
+import com.epam.taskgym.exception.MissingAttributes;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
