+++ src/test/java/com/epam/taskgym/TraineeServiceTest.java
@@ -3,6 +3,7 @@ package com.epam.taskgym;
 import com.epam.taskgym.dto.TraineeDTO;
 import com.epam.taskgym.entity.Trainee;
 import com.epam.taskgym.entity.User;
+import com.epam.taskgym.exception.*;
 import com.epam.taskgym.repository.TraineeRepository;
 import com.epam.taskgym.repository.TrainingRepository;
 import com.epam.taskgym.service.TraineeService;
