+++ src/test/java/com/epam/taskgym/TrainerServiceTest.java
@@ -8,10 +8,10 @@ import com.epam.taskgym.repository.TrainerRepository;
 import com.epam.taskgym.service.TrainerService;
 import com.epam.taskgym.service.TrainingTypeService;
 import com.epam.taskgym.service.UserService;
-import com.epam.taskgym.service.exception.FailAuthenticateException;
-import com.epam.taskgym.service.exception.InvalidPasswordException;
-import com.epam.taskgym.service.exception.MissingAttributes;
-import com.epam.taskgym.service.exception.NotFoundException;
+import com.epam.taskgym.exception.FailAuthenticateException;
+import com.epam.taskgym.exception.InvalidPasswordException;
+import com.epam.taskgym.exception.MissingAttributes;
+import com.epam.taskgym.exception.NotFoundException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
