+++ src/test/java/com/epam/taskgym/TrainingServiceTest.java
@@ -9,8 +9,8 @@ import com.epam.taskgym.service.TraineeService;
 import com.epam.taskgym.service.TrainerService;
 import com.epam.taskgym.service.TrainingService;
 import com.epam.taskgym.service.TrainingTypeService;
-import com.epam.taskgym.service.exception.BadRequestException;
-import com.epam.taskgym.service.exception.MissingAttributes;
+import com.epam.taskgym.exception.BadRequestException;
+import com.epam.taskgym.exception.MissingAttributes;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
