+++ src/test/java/com/epam/taskgym/TrainingTypeServiceTest.java
@@ -3,8 +3,8 @@ package com.epam.taskgym;
 import com.epam.taskgym.entity.TrainingType;
 import com.epam.taskgym.repository.TrainingTypeRepository;
 import com.epam.taskgym.service.TrainingTypeService;
-import com.epam.taskgym.service.exception.MissingAttributes;
-import com.epam.taskgym.service.exception.NotFoundException;
+import com.epam.taskgym.exception.MissingAttributes;
+import com.epam.taskgym.exception.NotFoundException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
