+++ src/test/java/com/epam/taskgym/UserServiceTest.java
@@ -3,7 +3,7 @@ package com.epam.taskgym;
 import com.epam.taskgym.entity.User;
 import com.epam.taskgym.repository.UserRepository;
 import com.epam.taskgym.service.UserService;
-import com.epam.taskgym.service.exception.MissingAttributes;
+import com.epam.taskgym.exception.MissingAttributes;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
